-Created new FXML file and new controller.
-Replaced "Integer.parseInt(currentAgeTextField.getText()" with "86" in
LifeExpectancyCalculatorController class to remove an error.
-(Copied code from LifeCalculatorController to
LifeExpectancyCalculatorController.)
