+++ src/application/LifeCalculatorView.fxml
